Added several optional new parameters to the videofig() function: fig, proc_func, cmap, winname
Changed scope of several variables so that multiple matplotlib figure windows can be open and playing image sequences simultaneously.  
Instead of accessing only the current figure via "plt.", we now use the figure handle. 

Added a default redraw_fn that is designed for image sequences and for calling a separate image processing function to get the results that are to be displayed.

Added examples of how to use it with scikit-image.